headタグ削除、metadataに置き換え途中
