+++ package.json
@@ -24,7 +24,7 @@
     "dayjs": "^1.11.7",
     "dotenv": "^16.0.3",
     "microcms-js-sdk": "^2.3.2",
-    "next": "^13.1.6",
+    "next": "^13.2.4",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "server-only": "^0.0.1",
