--- src/app/about/head.tsx
@@ -1,14 +1,0 @@
-import { DefaultHead } from '@/components/common/defaultHead'
-
-export default async function Head({}) {
-  return (
-    <>
-      <DefaultHead
-        title='ABOUT | HTMLGO'
-        description='このサイトについて'
-        image=''
-        url='https://htmlgo.site/about'
-      />
-    </>
-  )
-}
