--- src/app/articles/head.tsx
@@ -1,13 +1,0 @@
-import { DefaultHead } from '@/components/common/defaultHead'
-
-export default async function Head({}) {
-  return (
-    <>
-      <DefaultHead
-        title='ARTICLES | HTMLGO'
-        description='色々なブログの記事のフィードをまとめたものです'
-        url='https://htmlgo.site/article'
-      />
-    </>
-  )
-}
