--- src/app/contact/head.tsx
@@ -1,14 +1,0 @@
-import { DefaultHead } from '@/components/common/defaultHead'
-
-export default async function Head({}) {
-  return (
-    <>
-      <DefaultHead
-        title='CONTACT | HTMLGO'
-        description='お問い合わせはこちらからお願いします'
-        image=''
-        url='https://htmlgo.site/contact'
-      />
-    </>
-  )
-}
