--- src/app/works/head.tsx
@@ -1,13 +1,0 @@
-import { DefaultHead } from '@/components/common/defaultHead'
-
-export default async function Head({}) {
-  return (
-    <>
-      <DefaultHead
-        title='WORKS | HTMLGO'
-        description='今までの実績一覧です'
-        url='https://htmlgo.site/works'
-      />
-    </>
-  )
-}
