+++ src/const/config.ts
@@ -1,0 +1,2 @@
+export const baseURL = 'https://www.htmlgo.site'
+export const setBaseUrl = (relativePath: string) => baseURL + relativePath
